Refactored some responses and dtos
