+++ MundoBalloonApi.business/MundoBalloonApi.business.csproj
@@ -19,5 +19,8 @@
     <ItemGroup>
         <ProjectReference Include="..\MundoBalloonApi.infrastructure\MundoBalloonApi.infrastructure.csproj" />
     </ItemGroup>
+    <ItemGroup>
+      <Folder Include="DTOs\Collections" />
+    </ItemGroup>
 
 </Project>
